Add files via upload
Farm Management System is made using the JAVA (Swing , AWT ,etc)  for the Farmers for their daily use . It help to keep their record updates. It keeps their Crop details, Income and the Expenses details etc. They can see their data in the format of tables. It is easy to handle.  User can use it in a daily basis.